+++ src/components/dialog/dialog.js
@@ -3,6 +3,7 @@ import Bowser from 'bowser';
 import classNames from 'classnames';
 import { assign } from 'lodash';
 import PropTypes from 'prop-types';
+import Browser from './../../utils/helpers/browser';
 import Icon from './../icon';
 import Modal from './../modal';
 import Heading from './../heading';
@@ -70,12 +71,15 @@ class Dialog extends Modal {
 
   static defaultProps = {
     size: 'medium',
-    showCloseIcon: true
+    showCloseIcon: true,
+    ariaRole: 'dialog'
   }
 
   constructor(args) {
     super(args);
     this.componentTags = this.componentTags.bind(this);
+    this.onDialogBlur = this.onDialogBlur.bind(this);
+    this.onCloseIconBlur = this.onCloseIconBlur.bind(this);
   }
 
   /**
@@ -87,7 +91,29 @@ class Dialog extends Modal {
   componentDidMount() {
     if (this.props.open) {
       this.centerDialog();
+      this.focusDialog();
     }
+  }
+
+  /**
+   * Event handler to handle keyboard
+   * focus leaving the dialog element.
+   */
+  onDialogBlur(ev) { } // eslint-disable-line no-unused-vars
+
+  /**
+   * Event handler for when the close icon
+   * loses keyboard focus.
+   *
+   * As the close icon is the last element in the dialog
+   * source, when the keyboard focus leaves the close
+   * icon return the focus to the dialog itself.
+   *
+   * @return {Void}
+   */
+  onCloseIconBlur(ev) {
+    ev.preventDefault();
+    this.focusDialog();
   }
 
   /**
@@ -100,7 +126,8 @@ class Dialog extends Modal {
    */
   get onOpening() {
     this.centerDialog();
-    window.addEventListener('resize', this.centerDialog);
+    this.focusDialog();
+    this.window().addEventListener('resize', this.centerDialog);
   }
 
   /**
@@ -112,7 +139,17 @@ class Dialog extends Modal {
    * @return {Void}
    */
   get onClosing() {
-    window.removeEventListener('resize', this.centerDialog);
+    this.window().removeEventListener('resize', this.centerDialog);
+  }
+
+  /**
+   * Returns the current window
+   *
+   * @method window
+   * @return {Object}
+   */
+  window = () => {
+    return Browser.getWindow();
   }
 
   /**
@@ -124,8 +161,11 @@ class Dialog extends Modal {
   centerDialog = () => {
     const height = this._dialog.offsetHeight / 2,
         width = this._dialog.offsetWidth / 2;
-    let midPointY = (window.innerHeight / 2) + window.pageYOffset,
-        midPointX = (window.innerWidth / 2) + window.pageXOffset;
+
+    const win = this.window();
+
+    let midPointY = (win.innerHeight / 2) + win.pageYOffset,
+        midPointX = (win.innerWidth / 2) + win.pageXOffset;
 
     midPointY -= height;
     midPointX -= width;
@@ -133,7 +173,7 @@ class Dialog extends Modal {
     if (midPointY < 20) {
       midPointY = 20;
     } else if (Bowser.ios) {
-      midPointY -= window.pageYOffset;
+      midPointY -= win.pageYOffset;
     }
 
     if (midPointX < 20) {
@@ -142,6 +182,10 @@ class Dialog extends Modal {
 
     this._dialog.style.top = `${midPointY}px`;
     this._dialog.style.left = `${midPointX}px`;
+  }
+
+  focusDialog() {
+    this._dialog.focus();
   }
 
   /**
@@ -198,6 +242,8 @@ class Dialog extends Modal {
           data-element='close'
           onClick={ this.props.onCancel }
           type='close'
+          tabIndex='0'
+          onBlur={ this.onCloseIconBlur }
         />
       );
     }
@@ -219,11 +265,29 @@ class Dialog extends Modal {
    * @return {Object} JSX for dialog
    */
   get modalHTML() {
+    let dialogProps = {
+      className: this.dialogClasses,
+      tabIndex: 0
+    };
+
+    if (this.props.ariaRole) {
+      dialogProps['role'] = this.props.ariaRole;
+    }
+
+    if (this.props.title) {
+      dialogProps['aria-labelledby'] = 'carbon-dialog-title';
+    }
+
+    if (this.props.subtitle) {
+      dialogProps['aria-describedby'] = 'carbon-dialog-subtitle';
+    }
+
     return (
       <div
         ref={ (d) => { this._dialog = d; } }
-        className={ this.dialogClasses }
+        { ...dialogProps }
         { ...this.componentTags(this.props) }
+        onBlur={ this.onDialogBlur }
       >
         { this.closeIcon }