+++ CHANGELOG.md
@@ -141,10 +141,12 @@ A new helper object is available in `utils/helpers/text`. Currently it only cont
 
 ## Component Improvements
 
+* `Alert` now alerts itself to screen readers.
 * `Browser`: add a new method `setInputFocus` to focus on the input field of passed in ref but does not select text
-* `DialogFullScreen`: the content is now scrollable.
-* `DialogFullScreen` is now using the `Heading` component to render its title.
-* `DialogFullScreen`: the title can be passed as a child instead of using the title prop.
+* `DialogFullScreen`:
+  * the content is now scrollable.
+  * is now using the `Heading` component to render its title.
+  * the title can be passed as a child instead of using the title prop.
 * `Form` now has default `SaveButton` and `CancelButton` functional stateless components. The former can be overridden with a new prop of `customSaveButton`.
 * `InlineInputs` children are now wrapped by Columns by the component.
 * `Menu` has been updated to use a `<nav>` tag as its root element.